+++ src/main/java/by/gomselmash/aspiski/service/EditDevelopersService.java
@@ -2,24 +2,31 @@ package by.gomselmash.aspiski.service;
 
 import by.gomselmash.aspiski.model.Developer;
 import by.gomselmash.aspiski.repository.DeveloperRepository;
+import by.gomselmash.aspiski.repository.ProgramRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
+@Transactional
 public class EditDevelopersService {
     private final DeveloperRepository developerRepository;
+    private final ProgramRepository programRepository;
 
-    public EditDevelopersService(DeveloperRepository developerRepository) {
+    public EditDevelopersService(
+            DeveloperRepository developerRepository,
+            ProgramRepository programRepository
+    ) {
         this.developerRepository = developerRepository;
+        this.programRepository = programRepository;
     }
 
+    @Transactional(readOnly = true)
     public List<Developer> findAllDevelopersSorted() {
         return developerRepository.findAllByOrderByNameAsc();
     }
 
-    @Transactional
     public Developer saveDeveloper(Developer developer) {
         String name = developer.getName();
         if (developerRepository.existsByNameIgnoreCase(name)) {
@@ -28,7 +35,25 @@ public class EditDevelopersService {
         return developerRepository.save(developer);
     }
 
-    public void deleteDeveloperById(int id) {
+    public boolean updateDeveloper(Developer developer) {
+        int id = developer.getId();
+        if (developerRepository.existsById(id)) {
+            developerRepository.save(developer);
+            return true;
+        }
+        return false;
+    }
+
+    public boolean deleteDeveloperById(int id) {
+        boolean isExists = developerRepository.existsById(id);
+        if (!isExists) {
+            return false;
+        }
+        boolean isRelatedKey = programRepository.existsByDeveloper_Id(id);
+        if (isRelatedKey) {
+            return false;
+        }
         developerRepository.deleteById(id);
+        return true;
     }
 }
