+++ src/main/java/by/gomselmash/aspiski/service/EditMachineTypesService.java
@@ -1,6 +1,7 @@
 package by.gomselmash.aspiski.service;
 
 import by.gomselmash.aspiski.model.MachineType;
+import by.gomselmash.aspiski.repository.MachineRepository;
 import by.gomselmash.aspiski.repository.MachineTypeRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -8,18 +9,24 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Service
+@Transactional
 public class EditMachineTypesService {
     private final MachineTypeRepository machineTypeRepository;
+    private final MachineRepository machineRepository;
 
-    public EditMachineTypesService(MachineTypeRepository machineTypeRepository) {
+    public EditMachineTypesService(
+            MachineTypeRepository machineTypeRepository,
+            MachineRepository machineRepository
+    ) {
         this.machineTypeRepository = machineTypeRepository;
+        this.machineRepository = machineRepository;
     }
 
+    @Transactional(readOnly = true)
     public List<MachineType> findAllMachineTypesSorted() {
         return machineTypeRepository.findAllByOrderByNameAsc();
     }
 
-    @Transactional
     public MachineType saveMachineType(MachineType machineType) {
         String name = machineType.getName();
         if (machineTypeRepository.existsByNameIgnoreCase(name)) {
@@ -28,7 +35,25 @@ public class EditMachineTypesService {
         return machineTypeRepository.save(machineType);
     }
 
-    public void deleteMachineTypeById(int id) {
+    public Boolean updateMachineType(MachineType machineType) {
+        int id = machineType.getId();
+        if (machineTypeRepository.existsById(id)) {
+            machineTypeRepository.save(machineType);
+            return true;
+        }
+        return false;
+    }
+
+    public boolean deleteMachineTypeById(int id) {
+        boolean isExists = machineTypeRepository.existsById(id);
+        if (!isExists) {
+            return false;
+        }
+        boolean isRelatedKey = machineRepository.existsByMachineType_Id(id);
+        if (isRelatedKey) {
+            return false;
+        }
         machineTypeRepository.deleteById(id);
+        return true;
     }
 }
